[fastmanifest] fix conversion from flat manifest
Summary: When converting from a flat manifest, we inappropriately set the type of the root to IMPLICIT.  This meant that when we removed all the nodes in the tree, we would end up removing the root node, as it looked like just any other directory.

Test Plan: wrote a unit test to hit this scenario. it failed without the fix. passed with the fix.

Reviewers: wez, lcharignon

Reviewed By: lcharignon

Subscribers: mitrandir, mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D3289249

Tasks: 11145050

Signature: t1:3289249:1463063771:0e5d4f1bca296b1009c9ec6461f7aab85546fc8f