[fastmanifest] have debugfastmanifest consistently return a fastmanifest
Summary:
When fastmanifest.debugfastmanifest is set to True, we always return a fastmanifest.  If the cache is empty, we will convert a flat manifest into a fastmanifest and use that for all operations.

Same goal as D3247484.

Test Plan: pass some unit tests in the mercurial suite.  need to investigate the remaining issues.

Reviewers: #mercurial, lcharignon

Reviewed By: lcharignon

Subscribers: mitrandir, mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D3295434

Signature: t1:3295434:1463439017:0a6d20fa23fe409094e35a46eb195495455b8602