[fastmanifest] implement filter-copy
Summary:
Implement a copy mechanism that filters along the way.  For leaf nodes, it's pretty straightforward.

For implicit nodes, instead of allocating full-sized nodes on the arena, we allocate a temporary node, write the children there.  Once we finish, we determine whether or not we still need the implicit node (because we may end up with 0 children).  If it is, we create the permanent node and copy the entries over.  If a subtree is not fully copied over, we invalidate the checksum on the implicit node.

Test Plan: wrote a unit test that prunes the entire tree.  wrote a unit test that prunes a subtree.  verifies that the nodes and their checksums.

Reviewers: wez, lcharignon

Reviewed By: lcharignon

Subscribers: mitrandir, mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D3318895

Tasks: 10644559

Signature: t1:3318895:1463668333:e2bfee36ab4b324dc3a93dedbe9bb36459f2194b