[fastmanifest] fix creation of hybridmanifests when running with debugfastmanifest
Summary: We want to always present a fastmanifest, so we always return true for `_incache()`.  `_cachedmanifest()` is modified to extract the flat manifest when debugfastmanifest is set, and convert the extracted manifest.

Test Plan: down to 45 failures with this and the `filtercopy()` fixes. #10644559

Reviewers: lcharignon

Reviewed By: lcharignon

Subscribers: mitrandir, mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D3320099

Tasks: 11278537

Signature: t1:3320099:1463668296:3f0e6b37524d8a74f531bc418a5594e13994d2b2