sqldirstate: make shelve work with sqldirstate
Summary:
shelve is woking thanks to the hack where it abandons the transaction but
preserves the dirstate anyway by copying the dirstate to the side and restoring
it. We can do something that works as well and is way faster.

Test Plan:
ran tests
tried it on fbsource - it was nice and fast

Reviewers: #mercurial, ttung, simonfar, durham

Reviewed By: durham

Subscribers: simonfar, mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D3345280

Signature: t1:3345280:1464170069:cfb22b7034f8a8c44fd4284cd35ba17e85dabb99