Introduce InodeDispatcher::lookupInodeBase().
Summary:
This is an alternative to `InodeDispatcher::lookup()` that does not result
in `RequestData::get()` being called, which is important for requests that
do not originate from FUSE.

The `::lookup()` call takes the `InodeBase` and invokes its `getattr()` method,
which in the case of `TreeEntryFileInode::getattr()`, creates a `fusell::Dispatcher::Attr`
whose constructor calls the following:

```
Dispatcher::Attr::Attr() : timeout(1.0) {
  memset(&st, 0, sizeof(st));
  auto& req = RequestData::get();
  auto mount = req.getChannel().getMountPoint();
  st.st_uid = mount->getUid();
  st.st_gid = mount->getGid();
}
```

So the only reason this is done at all is to set `st_uid` and `st_gid`.

And then the only reason `Attr` is needed is to set the following fields on a `fuse_entry_param`:

```
attr
attr_timeout
entry_timeout
```

So it is possible that we can find a better way to streamline all of this, but this was
the easiest shortcut I could implement myself right now.

Reviewed By: wez

Differential Revision: D3351819

fbshipit-source-id: b095c085ee4a9b9a81438db093869fd0acf1f8ad