tweakdefaults: remove "head -n0" from test-tweakdefaults.t
Summary:
BSD version of head does not support -n0 and it looks like the only purpose of
using it here is to hide stdout output (and leave stderr).

Test Plan:
Ran test-tweakdefaults.t on both Linux and OSX, made sure that this particular
part of the test does not fail anymore.
That said, test-tweakdefaults.t still fails on OSX, but that requires more
investigation and my 5-minute look did not reveal the reason.

Reviewers: simonfar

Subscribers: mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D3378544