fix issues with reverse path iterators
Summary:
The reverse path iterators derived from the forward iterators, and overrode the
pre-increment operator.  Unfortunately they didn't override post-increment,
which meant that post increment actually advanced them forwards instead of
backwards.  (The base class's post-increment calls the base class's
pre-increment, since this method is not virtual.)  The reverse iterators also
weren't overriding the decrement operators added in D3366877, so decremening a
reverse iterator would move it in the wrong direction.

This fixes all of those issues by making the ComposedPathIterator template
smarter so it can correctly handle both forward and reverse iteration.
It also now contains the logic that the "begin" iteration for absolute paths
needs to start one character in.  ComposedPathIterator needs to be aware of
this in order to stop at the correct character when iterating backwards.

Reviewed By: bolinfest

Differential Revision: D3376727

fbshipit-source-id: c61c86c1a9233507f8297015439e416680aaa39b