add a kDirSeparator constant
Summary:
Define a constant for the path directory separator ('/').  We currently plan to
normalize all pathnames to use forward slash as a directory separator, even on
Windows.  This will simplify the bulk of the eden code.

Even though we don't plan to make this separator character configurable, it
still seems useful to define a symbolic constant for it.

Reviewed By: bolinfest

Differential Revision: D3377110

fbshipit-source-id: cf6b158824e0cb318cc94838618506e78d868c9f