improve start-up handling of hg_import_helper.py
Summary:
Look for hg_import_helper.py relative to the edenfs binary, rather than
relative to the current working directory.  Also check in a couple places, so
we can work in both normal deployment locations as well as inside a source
repository.

Additionally, update hg_import_helper.py to emit a response chunk as soon as it
starts.  This contains either message indicating that it has started
successfully, or an error message.  This allows us to propagate error
information back to the thrift caller if something goes wrong starting
hg_import_helper.py for a given repository.

Reviewed By: bolinfest

Differential Revision: D3447027

fbshipit-source-id: 4bffd1d03ab6475c09cb3c8385145555c7fc7361