sqldirstate: fast in-memory cache for batch operations
Summary:
To fix sqldirstate perf in case of iterating through whole
dirstate we are introducing the possibility to cache it in memory when
neccessary.

Currently we are doing it only in the case when dirstate.walk() is called.

Test Plan:
    ./run-sqlitedirstate-test.py --hg ~/clowncopter/ -j 30 -t 1200
    ~/clowncopter/tests/run-tests.py -j 30

Reviewers: #mercurial, durham, ttung, lcharignon

Reviewed By: lcharignon

Subscribers: lcharignon, mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D3472402

Signature: t1:3472402:1466807592:a28bc4662cc157547b8b747c6be98faae35c63b5