test_fetch_renames: avoid hash randomization bug in test
When hash randomization is enabled, the order in which revisions are
committed to the Mercurial repository can vary for revisions which
span branches. In order to work around this, we key the dictionary of
golden values for copy data using the suffix of the convert_revision
metadata rather than the Mercurial changelog index of the revision.