eden: fixup how we construct the fuse args struct
Summary:
I found this by running the tests with ASAN enabled; it was reporting leaks.

The issue is that fuse decides to append some args to the argument list in some
cases.  When it does this it tries to realloc the storage in the argv array.
This was not safe to do because that data was owned by the vector and may not
have been allocated directly by malloc.

This diff switches us to use the underlying functions in libfuse that append
arguments and then know how to safely free the result.

SCOPE_EXIT is used to ensure that this gets cleaned up when we return.

Reviewed By: bolinfest

Differential Revision: D3519999

fbshipit-source-id: ca12739f11be1fd9662063ec7515cec90ccf0d57